authinvite plugin - extract db helper module
